pyproject.toml
[build-system]
requires = ["setuptools>=61.0.0"] # プロジェクトをビルドするために必要なツール
build-backend = "setuptools.build_meta"

[project]
name = "aetherterm"
version = "0.0.1"
description = "A sleek web based terminal emulator"
readme = "README.md"
requires-python = ">=3.9"
authors = [
    {name = "Kazushige Takeuchi", email = "kaz@re-x.info"}
]
keywords = ["terminal", "web", "emulator", "websocket"]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Terminals"
]

dependencies = [
    "dependency-injector>=2.20.1",
    "python-socketio",
    "uvicorn",
    "fastapi",
    "python-multipart",
    "jinja2",
    "libsass",
    "pyopenssl",
    "click",
    "websockets",
    # LangChain関連依存関係
    "langchain>=0.1.0",
    "langchain-openai>=0.1.0",
    "langchain-anthropic>=0.1.0",
    "langchain-community>=0.0.20",
    "chromadb>=0.4.0",
    "faiss-cpu>=1.7.0",
    "redis>=5.0.0",
    "sqlalchemy>=2.0.0",
    "alembic>=1.13.0",
    "psycopg2-binary>=2.9.0",
    "asyncpg>=0.29.0",
    "sentence-transformers>=2.2.0",
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "tiktoken>=0.5.0",
]

[project.optional-dependencies]
themes = ["libsass"]
lint = ["pytest", "pytest-flake8", "pytest-isort"]
dev = ["ruff"]

[project.urls]
Homepage = "https://github.com/paradoxxxzero/aetherterm"
Repository = "https://github.com/paradoxxxzero/aetherterm"
Issues = "https://github.com/paradoxxxzero/aetherterm/issues"

[project.scripts]
aetherterm = "aetherterm.scripts.aetherterm:main"
aether = "aetherterm.scripts.aetherterm:main"
aetherterm-agentserver = "aetherterm.agentserver.main:main"
aetherterm-agentshell = "aetherterm.agentshell.main:main"
aetherterm-controlserver = "aetherterm.controlserver.main:main"
aetherterm-shell-monitor = "aetherterm.agentshell.pty_monitor.main:main"
aetherterm-dummy-ai = "aetherterm.agentshell.pty_monitor.dummy_ai_server:main"

[tool.setuptools]

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
aetherterm = [
    "agentserver/sass/*.sass",
    "agentserver/themes/*.*",
    "agentserver/themes/*/*.*", 
    "agentserver/themes/*/*/*.*",
    "agentserver/static/fonts/*",
    "agentserver/static/images/favicon.png",
    "agentserver/static/main.css",
    "agentserver/static/html-sanitizer.js",
    "agentserver/static/*.min.js",
    "agentserver/static/assets/*",
    "agentserver/templates/index.html",
    "agentserver/templates/motd.j2",
    "agentserver/bin/*",
    "agentserver/terminals/*.py",
    "aetherterm.conf.default"
]

[tool.pytest.ini_options]
flake8-ignore = [
    "*.py E731 E402",
    "aetherterm/bin/help.py E501"
]

[tool.ruff]
line-length = 100
[tool.ruff.lint]
select = ["E", "F", "W", "I"] # エラーコードの選択

[tool.wheel]
universal = true

[dependency-groups]
dev = [
    "hypercorn>=0.17.3",
]

